Modify json number and tests
Make modification for method to work with fractional numbers, and test it.

+++ Json.Validator/JsonNumber.cs
@@ -11,7 +11,7 @@ namespace Json
                 return false;
             }
 
-            return int.TryParse(input, out _);
+            return double.TryParse(input, out _);
         }
 
         static bool StartsWithZeroOrIsNullOrEmpty(string input)
